requirements.txt
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
yfinance==0.2.18                 # For Yahoo Finance data fetching
alpha_vantage==2.3.1            # For Alpha Vantage API access
python-dotenv==1.0.0            # For .env file reading
gspread==5.7.1                  # Google Sheets API client
google-auth==2.21.0             # Authentication library for Google API
oauth2client==4.1.3             # Older GSheet auth utils, sometimes required
requests==2.31.0                # HTTP requests
schedule==1.1.0                 # For running scheduled jobs (optional, if used)
streamlit==1.24.1               # For frontend/dashboard
python-telegram-bot==20.3       # For Telegram bot integration (if implemented)
